fix(wasm): @form vec/hashmap/ring_buffer use target-pointer-width size_t (no more trap)
@form(vec) / @form(hashmap) / @form(ring_buffer) compiled for wasm but
trapped at runtime (RuntimeError: signature_mismatch:lotus_vec_push). Same
class as the bus-codec fix (69925dc): the collection runtime primitives'
`size_t` params still used native width.

Root cause: codegen declared the `lotus_vec_*` / `lotus_hashmap_*` /
`lotus_ring_buffer_*` `size_t` params (elem_size / key_size / value_size /
fixed_cap / cap) as i64, but the C runtime's `size_t` is i32 on wasm32 —
so the wasm `call`'s signature mismatched the definition (wasm-ld even
warned: "lotus_vec_push defined as (i32, i64, i32) vs (i32, i32, i32)").
Native matched (size_t == i64), so it never surfaced there.

Fix:
- builtins.rs: the `size_t` params now use `usize_type()` (i64 native /
  i32 wasm32). The `int64_t` index/len params and the `int` key-type tag
  are unchanged (i64 / i32 on both).
- form/mod.rs + locus/instantiation.rs: call sites narrow the `size_of()`
  size value with `size_to_usize` (no-op native), and build cap/fixed_cap
  at the usize width.
- lotus_arena.c: `lotus_ring_buffer_len` returned `size_t` (i32 wasm),
  mismatching the i64 `len()` accessor — changed to `int64_t` to match
  lotus_vec_len / lotus_hashmap_len. The count is bounded by cap.

Verified: vwasm.mjs/hwasm.mjs/rb.mjs all run; a full vec+hashmap+
ring_buffer program prints the right values under node. New
`wasm_form_collections_run` regression (the 2nd target-wasm run fixture,
after the bus one). Native unaffected (size_to_usize is a no-op when
usize==i64) — form_vec/hashmap/ring_buffer codegen suites green.
spec/docs (webassembly.md) note @form collections run under wasm.

+++ crates/hale-codegen/runtime/lotus_arena.c
@@ -4113,9 +4113,13 @@ int lotus_ring_buffer_pop(void *rb_ptr, void *out) {
     return 1;
 }
 
-size_t lotus_ring_buffer_len(void *rb_ptr) {
+/* Returns int64_t (a Hale `Int`) to match lotus_vec_len /
+ * lotus_hashmap_len — NOT size_t, whose target-pointer width (i32 on
+ * wasm32) would mismatch the i64 codegen declaration of the `len()`
+ * accessor. The count is bounded by `cap`, so it fits int64_t trivially. */
+int64_t lotus_ring_buffer_len(void *rb_ptr) {
     if (!rb_ptr) return 0;
-    return ((lotus_ring_buffer_t *)rb_ptr)->len;
+    return (int64_t)((lotus_ring_buffer_t *)rb_ptr)->len;
 }
 
 int lotus_ring_buffer_is_full(void *rb_ptr) {
